+++ src/liblsquic/lsquic_parse_gquic_be.h
@@ -82,8 +82,8 @@ int
 gquic_be_gen_ping_frame (unsigned char *buf, int buf_len);
 
 int
-gquic_be_gen_connect_close_frame (unsigned char *buf, int buf_len, uint32_t error_code,
-                            const char *reason, int reason_len);
+gquic_be_gen_connect_close_frame (unsigned char *buf, size_t buf_len,
+                    uint32_t error_code, const char *reason, int reason_len);
 
 int
 gquic_be_parse_connect_close_frame (const unsigned char *buf, size_t buf_len,
